+++ packages/manager/modules/billing-components/src/components/commitment-duration/commitment-duration.component.js
@@ -3,9 +3,11 @@ import template from './commitment-duration.html';
 
 export default {
   bindings: {
+    preselectDefault: '<?',
     defaultPrice: '<?',
     duration: '=',
     pricings: '<',
+    selectedQuantity: '<',
     fromCatalog: '<?',
     onDurationChange: '&',
   },