+++ packages/manager/modules/billing-components/src/components/commitment-pricing-mode/commitment-pricing-mode.component.js
@@ -3,6 +3,7 @@ import template from './commitment-pricing-mode.html';
 
 export default {
   bindings: {
+    isHideableDiscount: '<',
     pricingMode: '=',
     pricingModes: '<',
     duration: '<',