+++ packages/manager/modules/netapp/src/dashboard/snapshot-policies/component.js
@@ -7,6 +7,7 @@ export default {
     getSnapshotPolicies: '<',
     serviceName: '<',
     snapshotPolicies: '<',
+    trackClick: '<',
   },
   controller,
   template,