+++ packages/manager/modules/ng-layout-helpers/src/list/list-layout.component.js
@@ -16,6 +16,8 @@ export default {
     getServiceNameLink: '<?',
     options: '<?',
     formatters: '<?',
+    topbarOptions: '<?',
+    customizeColumnsMap: '<?',
   },
   controller,
   template,