+++ packages/manager/modules/ng-layout-helpers/src/list/list-layout.utils.js
@@ -87,12 +87,21 @@ const getTypeOptions = (schema, type, column, columnType) => {
 export const getSorting = ({ sort, sortOrder }, property) =>
   sort === property ? sortOrder.toLowerCase() : true;
 
-export const parseConfig = (columns, model, schema, sorting, tracking) =>
+export const parseConfig = (
+  columns,
+  model,
+  schema,
+  sorting,
+  tracking,
+  customizeColumnsMap = {},
+) =>
   map(columns, (column) => {
     const propertyType = get(model.properties[column.property], 'type');
     const type = getMatchingType(propertyType);
+    const customizeColumn = customizeColumnsMap[column.property] || {};
+
     return {
-      title: column.label,
+      title: customizeColumn?.title || column.label,
       property: column.property,
       ...(column.serviceLink ? { template: getLink(column, tracking) } : {}),
       ...(column.format ? { template: generateTemplate(column) } : {}),
@@ -280,10 +289,13 @@ export const stateResolves = {
     sortOrder,
   ) => {
     let serviceNameTracker;
+    let customizeColumnsMap;
     try {
       serviceNameTracker = $transition$.injector().get('serviceNameTracker');
+      customizeColumnsMap = $transition$.injector().get('customizeColumnsMap');
     } catch (error) {
       serviceNameTracker = null;
+      customizeColumnsMap = {};
     }
 
     return parseConfig(
@@ -292,6 +304,7 @@ export const stateResolves = {
       schema,
       { sort, sortOrder },
       serviceNameTracker,
+      customizeColumnsMap,
     );
   },
   formatters: /* @ngInject */ (configuration) => {