+++ packages/manager/apps/netapp/CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.1.6](https://github.com/ovh/manager/compare/@ovh-ux/manager-netapp-app@0.1.5...@ovh-ux/manager-netapp-app@0.1.6) (2022-03-02)
+
+
+### Bug Fixes
+
+* **netapp:** use cucBytes to get right value and unit ([#6251](https://github.com/ovh/manager/issues/6251)) ([d07231b](https://github.com/ovh/manager/commit/d07231bcc699ed644b765ede9e92b795095afd1a))
+* **netapp:** use cucBytes to get right value and unit ([#6251](https://github.com/ovh/manager/issues/6251)) ([fb654b7](https://github.com/ovh/manager/commit/fb654b78d24b792201c2b58169d1cc2d37c6be08))
+
+
+
 ## [0.1.5](https://github.com/ovh/manager/compare/@ovh-ux/manager-netapp-app@0.1.4...@ovh-ux/manager-netapp-app@0.1.5) (2022-01-25)